api-gateway/requirements.txt
flask==2.3.2
prometheus-flask-exporter==0.22.4
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-flask